Add static member detection to handle default values for registerNode
We add the FullOpinionatedLocator which provides all the possible static
members for the arguments of registerNode and PartialOpinionatedNode
which only provides the minimum set (the ones for which there is no
default argument.)

Then we add static member detection code using the SFINAE principle
(Substitution Failure Is Not An Error.) We test for the presence of the
nodeType and classification member variables and use their values if
they are there and the expected defaults if they are not.

We update the "opinionated-example-plugin" code to register the
FullOpinionatedLocator and PartialOpinionatedNode using this machinery.